// web-ui/src/Models/Interfaces/Rules.tsx
export interface Rules {
    maxTeamSize: number
    maxPlayersPerCollegeTeam: number
    pointsPerCleanSheet: number
    pointsPerDefenderGoal: number
    pointsPerMidfielderGoal: number
    pointsPerAttackerGoal: number
    pointsPerAssists: number
    pointsPerRedCard: number
    pointsPerYellowCard: number
    manOfTheMatchBonus: number
    initialBudget: number
  }